requirements.txt
streamlit>=1.32
pandas
gspread
google-auth
fpdf
openpyxl
